+++ config/settings.yaml
@@ -14,6 +14,7 @@ urls:
   alerts: "https://the-internet.herokuapp.com/javascript_alerts"
   context_menu: "https://the-internet.herokuapp.com/context_menu"
   horizontal_slider: "https://the-internet.herokuapp.com/horizontal_slider"
+  hovers: "http://the-internet.herokuapp.com/hovers"
 
 # Учетные данные (можно вынести в secrets)
 credentials:
